requirements.txt
aiohttp
beautifulsoup4
cloudscraper
nest_asyncio
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client